+++ idea/src/org/jetbrains/kotlin/idea/refactoring/KotlinRefactoringSupportProvider.kt
@@ -23,6 +23,7 @@ import com.intellij.psi.PsiNameIdentifierOwner
 import com.intellij.refactoring.RefactoringActionHandler
 import org.jetbrains.kotlin.idea.refactoring.changeSignature.KotlinChangeSignatureHandler
 import org.jetbrains.kotlin.idea.refactoring.introduce.extractFunction.ExtractKotlinFunctionHandler
+import org.jetbrains.kotlin.idea.refactoring.introduce.extractClass.KotlinExtractSuperclassHandler
 import org.jetbrains.kotlin.idea.refactoring.introduce.introduceParameter.KotlinIntroduceLambdaParameterHandler
 import org.jetbrains.kotlin.idea.refactoring.introduce.introduceParameter.KotlinIntroduceParameterHandler
 import org.jetbrains.kotlin.idea.refactoring.introduce.introduceProperty.KotlinIntroducePropertyHandler
@@ -78,6 +79,8 @@ class KotlinRefactoringSupportProvider : RefactoringSupportProvider() {
     override fun getPullUpHandler() = KotlinPullUpHandler()
 
     override fun getPushDownHandler() = KotlinPushDownHandler()
+
+    override fun getExtractSuperClassHandler() = KotlinExtractSuperclassHandler
 }
 
 class KotlinVetoRenameCondition: Condition<PsiElement> {