+++ idea/src/org/jetbrains/kotlin/idea/refactoring/introduce/introduceTypeAlias/introduceTypeAliasImpl.kt
@@ -27,6 +27,7 @@ import org.jetbrains.kotlin.idea.core.CollectingNameValidator
 import org.jetbrains.kotlin.idea.core.KotlinNameSuggester
 import org.jetbrains.kotlin.idea.core.compareDescriptors
 import org.jetbrains.kotlin.idea.core.quoteIfNeeded
+import org.jetbrains.kotlin.idea.refactoring.introduce.insertDeclaration
 import org.jetbrains.kotlin.idea.util.getResolutionScope
 import org.jetbrains.kotlin.idea.util.psi.patternMatching.KotlinPsiRange
 import org.jetbrains.kotlin.idea.util.psi.patternMatching.KotlinPsiUnifier
@@ -220,22 +221,6 @@ fun IntroduceTypeAliasDescriptor.generateTypeAlias(previewOnly: Boolean = false)
         }
     }
 
-    fun insertDeclaration(): KtTypeAlias {
-        val targetParent = originalData.targetSibling.parent
-
-        val anchorCandidates = SmartList<PsiElement>()
-        anchorCandidates.add(targetSibling)
-        if (targetSibling is KtEnumEntry) {
-            anchorCandidates.add(targetSibling.siblings().last { it is KtEnumEntry })
-        }
-
-        val anchor = anchorCandidates.minBy { it.startOffset }!!.parentsWithSelf.first { it.parent == targetParent }
-        val targetContainer = anchor.parent!!
-        return (targetContainer.addBefore(typeAlias, anchor) as KtTypeAlias).apply {
-            targetContainer.addBefore(psiFactory.createWhiteSpace("\n\n"), anchor)
-        }
-    }
-
     return if (previewOnly) {
         introduceTypeParameters()
         typeAlias
@@ -243,6 +228,6 @@ fun IntroduceTypeAliasDescriptor.generateTypeAlias(previewOnly: Boolean = false)
     else {
         replaceUsage()
         introduceTypeParameters()
-        insertDeclaration()
+        insertDeclaration(typeAlias, originalData.targetSibling)
     }
 }