+++ idea/src/org/jetbrains/kotlin/idea/refactoring/memberInfo/KotlinMemberInfo.kt
@@ -23,6 +23,7 @@ import com.intellij.refactoring.util.classMembers.MemberInfo
 import org.jetbrains.kotlin.asJava.LightClassUtil
 import org.jetbrains.kotlin.asJava.getRepresentativeLightMethod
 import org.jetbrains.kotlin.asJava.toLightClass
+import org.jetbrains.kotlin.asJava.unwrapped
 import org.jetbrains.kotlin.descriptors.CallableMemberDescriptor
 import org.jetbrains.kotlin.descriptors.MemberDescriptor
 import org.jetbrains.kotlin.descriptors.Modality
@@ -78,4 +79,11 @@ fun KotlinMemberInfo.toJavaMemberInfo(): MemberInfo? {
     val info = MemberInfo(psiMember ?: return null, isSuperClass, null)
     info.isToAbstract = isToAbstract
     return info
+}
+
+fun MemberInfo.toKotlinMemberInfo(): KotlinMemberInfo? {
+    val declaration = member.unwrapped as? KtNamedDeclaration ?: return null
+    return KotlinMemberInfo(declaration, declaration is KtClass && overrides != null).apply {
+        this.isToAbstract = this@toKotlinMemberInfo.isToAbstract
+    }
 }