+++ idea/testData/refactoring/extractSuperclass/addSuperclassNoSecondaryConstructors.kt
@@ -1,0 +1,10 @@
+// NAME: X
+interface T {}
+
+// SIBLING:
+class <caret>A : T {
+    // INFO: {checked: "true"}
+    fun foo() {
+
+    }
+}